Refactor task logs API
