+++ web/src/application/state/tasks/logs/reducer.js
