+++ web/src/application/state/tasks/root/initialState.js
@@ -1,5 +1,6 @@
 import cacheInitialState from "../cache/initialState";
 import queriesInitialState from "../queries/initialState";
+import logsInitialState from "../logs/initialState";
 
 /**
  * Initial State for background task logic.
@@ -13,6 +14,10 @@ const initialState = {
    * Cached task queries.
    */
   queries: queriesInitialState,
+  /**
+   * Task logs state.
+   */
+  logs: logsInitialState,
 };
 
 export default initialState;
